Finish fixing doctests
