+++ speechbrain/lm/arpa.py
@@ -312,7 +312,7 @@ def arpa_to_fst(
     ...     + "<s> 3\n"
     ...     + "#0 4")  # Ends whole file
     >>> out = getfixture('tmpdir').join("bigram.txt.fst")
-    >>> arpa_to_fst(vocav, arpa_file, out, 2)
+    >>> arpa_to_fst(vocav, arpa_file, out, 2) # doctest: +SKIP
     """
     try:
         from kaldilm.arpa2fst import arpa2fst
